python-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
slowapi>=0.1.9
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.4
hypothesis>=6.98.3
supabase>=2.0.0
postgrest>=0.16.0
psycopg2-binary>=2.9.9
email-validator>=2.0.0
